+++ my-app/src/components/main/searchBar/SearchBar.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useDispatch } from 'react-redux';
 import { filterGenreMoviesAsync, setCurrentPage, setMoviesKeyword } from '../../../redux/actions';
-import { getMoviesAPI } from '../../../redux/asyncActions';
+import { getMoviesAPI } from '../../../redux/asyncActionsThunks';
 import { useTypedSelector } from '../../../hooks/useTypedSelector';
 import './SearchBar.scss';
